requirements.txt
pytube>=12.1.0
nltk>=3.8.1
notion-client>=2.0.0
requests>=2.28.2
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.0
groq>=0.4.0
google-api-python-client>=2.79.0